e2e: add test image and helpers
In order to run cpu management tests, we need nontrivial
hardware discovery, more than the node status reports,
and some introspection in pods to learn their CPU assignment.

We can implement these using busybox and shell commands,
but past attempts doing that (e.g. kubernetes e2e tests)
proved awkward, so it should be the last resort;
furthermore we already have all the code we need, so
we just need to repackage it into a helper image,
which is much more handy and a little, fair price to pay.

This commit add these entrypoints and the test image.

Signed-off-by: Francesco Romani <fromani@redhat.com>