ci: consolidate setup steps
consolidate the CI kind setup step in just one target
vs the multiple fragmented target we had.
It's now much easier to setup a temporary cluster, at
cost of some duplication. We will review the duplication
later on, but the cost is pretty low anyway.

Note: we now use a specific Kind configuration for CI,
using just 1 worker node. For the CI needs, this is
the minimal configuration.

Signed-off-by: Francesco Romani <fromani@redhat.com>