e2e: test for exclusive CPU assignment
add initial test to automatically validate the exclusive
cpu assignment.

The test wants to check that exclusively-allocated CPUs
are not accessible to pods not requiring exclusive CPUs,
regardless if they are created before or after the pod
with resource claims.

At the moment, we depend on logs to learn the CPU assignment;
to refresh the information, we just let the container exit
and be restarted. This is a simple pattern well supported
by the platform which should cause no issue in the CI,
provided the tests are done in a Eventually/Consistently loop,
which is good practice anyway.

In the future, we may want add functionality to exec
straight into the tester pods to query the current cgroup
assignment, but this requires extra code and the
benefits are not evident yet.

Signed-off-by: Francesco Romani <fromani@redhat.com>