bugfix: cdi: write the CDI spec atomically
POSIX.1-2017 mandates that rename() is an atomic operation.
By writing to a temporary file and renaming it over the destination,
we guarantee that concurrent readers see either the old file or the
new file, but never a partially written or corrupted state.
Updating in-place introduces a race condition where readers may
observe an empty or incomplete file.
Updating in-place also may cause a corrupted file to be created
if the writing process crashes mid-write. Using the pattern proposed
in this commit, and widely known in literature, the file is always
complete and cannot be corrupted by a partial write.

Signed-off-by: Francesco Romani <fromani@redhat.com>